==> Notebooks/1.1_cleaning.ipynb <==
## Import Required Libraries

import pandas as pd
import numpy as np

df = pd.read_csv("../Data/Raw/BankChurners.csv")


df.head()

df.shape
df.head()
df.info()
df.describe()

## Remove useless columns

df.drop(columns=[
    'Naive_Bayes_Classifier_Attrition_Flag_Card_Category_Contacts_Count_12_mon_Dependent_count_Education_Level_Months_Inactive_12_mon_1',
    'Naive_Bayes_Classifier_Attrition_Flag_Card_Category_Contacts_Count_12_mon_Dependent_count_Education_Level_Months_Inactive_12_mon_2'
], inplace=True)

## Clean target variable (Attrition_Flag)

df['Attrition_Flag'] = df['Attrition_Flag'].map({
    'Existing Customer': 0,
    'Attrited Customer': 1
})

## Gender (Binary categorical)

df['Gender'] = df['Gender'].map({
    'M': 1,
    'F': 0
})

## Income_Category (ORDINAL + RANGES)

income_map = {
    'Less than $40K': 30000,
    '$40K - $60K': 50000,
    '$60K - $80K': 70000,
    '$80K - $120K': 100000,
    '$120K +': 130000,
    'Unknown': np.nan
}

df['Income_Value'] = df['Income_Category'].map(income_map)



df['Income_Value'].fillna(df['Income_Value'].median(), inplace=True)

df.drop(columns=['Income_Category'], inplace=True)

## Education_Level (Ordinal)

education_map = {
    'Uneducated': 1,
    'High School': 2,
    'College': 3,
    'Graduate': 4,
    'Post-Graduate': 5,
    'Doctorate': 6,
    'Unknown': np.nan
}
df['Education_Level'] = df['Education_Level'].map(education_map)


df['Education_Level'].fillna(df['Education_Level'].median(), inplace=True)

## Marital_Status (Nominal)

df['Marital_Status'] = df['Marital_Status'].fillna('Unknown')
df = pd.get_dummies(df, columns=['Marital_Status'], drop_first=True)

## Card_Category (Nominal)

df = pd.get_dummies(df, columns=['Card_Category'], drop_first=True)

## Handle remaining “Unknown” values

for col in df.select_dtypes(include='object').columns:
    df[col] = df[col].replace('Unknown', np.nan)

df.fillna(df.median(numeric_only=True), inplace=True)

## Check duplicates

df.duplicated().sum()
df.drop_duplicates(inplace=True)

## Final validation

df.info()
df.isna().sum()

## Drop "CLIENTNUM" column

df.drop(columns=["CLIENTNUM"], inplace=True)

df.info()

bool_cols = df.select_dtypes(include='bool').columns
df[bool_cols] = df[bool_cols].astype(int)


df.head()

## Save cleaned dataset

df.to_csv("../data/cleaned/BankChurners_cleaned.csv", index=False)
