==> Notebooks/4_features.ipynb <==
import pandas as pd
import numpy as np

df = pd.read_csv("../data/cleaned/Cleaned_data.csv")
df.head()

df.columns

# Feature_1
df["Activity_Ratio"] = df["Total_Trans_Amt"] / df["Total_Trans_Ct"]

# Feature_2
df["Engagement_Score"] = (
    df["Months_on_book"]
    + df["Contacts_Count_12_mon"]
    + df["Total_Trans_Ct"]
)

# Feature_3
median_credit = df["Credit_Limit"].median()
df["High_Credit_Flag"] = (df["Credit_Limit"] > median_credit).astype(int)

# Feature_4
df["Trans_Freq_Group"] = pd.qcut(
    df["Total_Trans_Ct"],
    q=3,
    labels=["Low", "Medium", "High"]
)


# Feature_5
def tenure_group(months):
    if months <= 24:
        return "New"
    elif months <= 48:
        return "Medium"
    else:
        return "Loyal"

df["Tenure_Group"] = df["Months_on_book"].apply(tenure_group)

df.to_csv("../data/Engineered/BankChurners_Engineered.csv",index=False)
